move all SNI calls to SNI packaged client
