+++ apps/web/lib/sni.ts
@@ -1,7 +1,9 @@
 'use client'
 
+import SNIClient from '@repo/sni'
 import useSWR from 'swr'
-import { currentScreen, listDevices, readDirectory } from './api'
+
+export const SNI = new SNIClient()
 
 const fetcher = async (key: string | string[] | null) => {
   if (!key) {
@@ -9,19 +11,19 @@ const fetcher = async (key: string | string[] | null) => {
   }
   switch (true) {
     case key === 'devices':
-      return await listDevices()
+      return await SNI.listDevices()
     case key.includes('readDirectory'):
       if (!key[1] || !key[2]) {
         throw new Error('Invalid URI or path')
       }
       return {
-        data: await readDirectory(key[2], key[1]),
+        data: await SNI.readDirectory(key[2], key[1]),
       }
     case key.includes('currentScreen'):
       if (!key[1]) {
         throw new Error('Invalid URI')
       }
-      return { data: await currentScreen(key[1]) }
+      return { data: await SNI.currentScreen(key[1]) }
     default:
       return null
   }
