--- apps/web/lib/sni/index.ts
@@ -1,3 +1,0 @@
-export * as sni from './sni'
-export * as sniClient from './client'
-export { useSNI } from './useSNI'
