+++ packages/sni/src/index.ts
@@ -42,29 +42,63 @@ const mapCapabilities = (input: number[]) =>
     )
   })
 
-  const getFolders = (input: any[], path: string) => {
-    const ignoreFolders = ['System Volume Information']
-    const folders = input
-      .filter(
-        (entry: any) =>
-          !entry.name.startsWith('.') && !ignoreFolders.includes(entry.name),
-      )
-      .sort((a, b) => a.name.toLowerCase().localeCompare(b.name.toLowerCase()))
-    return folders.map((entry) => ({
-      ...entry,
-      path: [path, entry.name].join('/').replace('//', '/'),
-    }))
+const fetchFields = () => {
+  const clone = { ...SNI.Field }
+  // remove all keys that are numbers
+  Object.keys(clone).forEach((key: string | number) => {
+    if (!isNaN(Number(key))) {
+      delete clone[key as keyof typeof clone]
+    }
+  })
+  return clone
+}
+
+const FIELDS = fetchFields()
+
+const mapFields = (input: string[]) => {
+  return input.map((field) => {
+    return Object.keys(FIELDS).find(
+      (key) => FIELDS[key as keyof typeof FIELDS] === field,
+    )
+  })
+}
+
+const getFolders = (input: any[], path: string) => {
+  const ignoreFolders = ['System Volume Information']
+  const folders = input
+    .filter(
+      (entry: any) =>
+        !entry.name.startsWith('.') && !ignoreFolders.includes(entry.name),
+    )
+    .sort((a, b) => a.name.toLowerCase().localeCompare(b.name.toLowerCase()))
+  return folders.map((entry) => ({
+    ...entry,
+    path: [path, entry.name].join('/').replace('//', '/'),
+  }))
+}
+
+const getFiles = (input: any[], path: string) => {
+  const files = input
+    .filter((entry: any) => !entry.name.startsWith('.'))
+    .sort((a, b) => a.name.toLowerCase().localeCompare(b.name.toLowerCase()))
+  return files.map((entry) => ({
+    ...entry,
+    path: [path, entry.name].join('/').replace('//', '/'),
+  }))
+}
+
+const validatePath = (path: string) => {
+  try {
+    if (path.length === 0) {
+      throw new Error('No path provided')
+    }
+    return true
+  } catch (err: unknown) {
+    const error = err as Error
+    console.error(`Could not validate path '${path}':`, error.message)
+    throw error
   }
-  
-  const getFiles = (input: any[], path: string) => {
-    const files = input
-      .filter((entry: any) => !entry.name.startsWith('.'))
-      .sort((a, b) => a.name.toLowerCase().localeCompare(b.name.toLowerCase()))
-    return files.map((entry) => ({
-      ...entry,
-      path: [path, entry.name].join('/').replace('//', '/'),
-    }))
-  }
+}
 
 class SNIClient {
   transport: GrpcWebFetchTransport
@@ -73,6 +107,49 @@ class SNIClient {
   constructor() {
     this.transport = setupTransport()
     this.clients = setupClients(this.transport)
+  }
+
+  async currentScreen (uri: string) {
+    try {
+      const req = await this.getFields(uri, ['RomFileName'])
+      if (req.values[0] !== '/sd2snes/m3nu.bin') {
+        return 'game'
+      }
+      return 'menu'
+    } catch (err: unknown) {
+      const error = err as Error
+      console.error('currentScreen', error.message)
+      return 'menu'
+    }
+  }
+
+  async bootFile (uri: string, path: string) {
+    validatePath(path)
+  
+    const req = SNI.BootFileRequest.create({ uri, path })
+    await this.clients.DeviceFilesystem.bootFile(req)
+    return path
+  }
+  
+  async deleteFile (uri: string, path: string) {
+    validatePath(path)
+  
+    const req = SNI.RemoveFileRequest.create({ uri, path })
+    await this.clients.DeviceFilesystem.removeFile(req)
+    return path
+  }
+
+  async getFields (uri: string, inputFields: string[]) {
+    const fields = inputFields.filter(
+      (field) => FIELDS[field as keyof typeof FIELDS],
+    )
+    if (fields.length === 0) {
+      throw new Error('No valid fields provided')
+    }
+    // TODO: Read fields from here
+    const req = SNI.FieldsRequest.create({ uri, fields: [SNI.Field.RomFileName] })
+    const call = await this.clients.DeviceInfo.fetchFields(req)
+    return call.response
   }
 
   async listDevices(kinds?: string[]) {
@@ -111,7 +188,7 @@ class SNIClient {
 
   async readDirectory (uri: string, path: string) {
     const req = SNI.ReadDirectoryRequest.create({ path, uri })
-    const call = await this.clients.readDirectory(req)
+    const call = await this.clients.DeviceFilesystem.readDirectory(req)
     const data = call.response.entries
     const rawFiles = data.filter((entry: any) => entry.type === 1)
     const rawFolders = data.filter((entry: any) => entry.type === 0)
